requirements.txt
faiss-cpu
sentence-transformers
streamlit
numpy
pandas
ollama
beautifulsoup4
requests
tqdm
